Move UserResponse class to a seperate file. Orginise imports.
