+++ src/resolvers/object-types/UserResponse.ts
@@ -1,0 +1,20 @@
+import { User } from "src/entities/User"
+import { ObjectType, Field } from "type-graphql"
+
+@ObjectType()
+class FieldError {
+   @Field()
+   field: string
+
+   @Field()
+   message: string
+}
+
+@ObjectType()
+export class UserResponse {
+   @Field(() => [FieldError], { nullable: true })
+   errors?: FieldError[]
+
+   @Field(() => User, { nullable: true })
+   user?: User
+}
