+++ src/types/types.ts
@@ -1,7 +1,7 @@
-import { EntityManager, IDatabaseDriver, Connection } from "@mikro-orm/core";
+import { Connection, EntityManager, IDatabaseDriver } from "@mikro-orm/core";
+import { Request, Response } from 'express';
 // import { ObjectId } from "@mikro-orm/mongodb";
 import { Redis } from "ioredis";
-import { Request, Response } from 'express'
 
 export type OrmContext = {
    em: EntityManager<IDatabaseDriver<Connection>>
